Refactor to a Server-Client scheme
This is a major refactor that fundamentally changes how nixbox works.
Instead of spawning a process that `unshare`s and then lives as a dummy
whose namespaces other processes can inherit, we will instead use a more
straight-forward solution of having a server that manages children
inside the namespaces environment. Eventually this will be a PID 1
service that can spawn and reap children.

Additionally, the user's home is not mounted in the namespaced
environment. Instead, it maintains a separate home directory. This
allows for eg. `~/.bashrc` to not interfere with the user's own
configuration.

Most importantly, there is now a host/guest distinction. Guest lives in
the NixOS enviornment while Host lives in the user's default
environment. Both are exposed as D-Bus services. `nixbox run [..]` will
now use D-Bus to run a NixOS command, while `nixbox run --host [..]`
will eventually allow NixOS programs to escape out of NixOS and into the
user's OS. Specifically, this is to allow services like `podman`,
`ssh-agent` and `gpg-agent`. If `gpg-agent` is not running, then it'll
automatically start the NixOS version. If the keys are in the user's
host, then these will not be available inside of NixOS. However, if
`gpg-agent` was started on the host, then inside of NixOS things will
encrypt/decrypt fine. Therefore, eventually we'll likely override the
`nixpkgs` versions of programs like `gpg-agent` to make them run `nixbox
run --host gpg-agent` instead.